Update sosa.ttl
I went into more details what a Procedure is and what it can be used for.

I removed all subclasses of Activity (such as Actuation) but not Activity itself. The description of Activity now names the previous subclasses as examples, e.g., a sampling activity. I changed all the comments that referred to these classes while making sure that the descriptions clarify that different procedures can be set up for different activities. Summing up, we have something like a soft hierarchy now, i.e., no axiomatization but natural descriptions for which kinds of activities and procedures can exist/ be useful.

I removed the strong link between Observation and Activity. We have to agree first whether an Observation is an Information Object or not. This is basically reverting the changes to the original SSN view *without* implying that this is the right way to go.

I removed Device and Platform but left Observer. Before somebody panics, there is actually a good reason for doing so based on yesterday's emails. We discussed before that we need something like SensingDevice for the thing that carries one or multiple sensors. We also discussed that this should not be Device because humans are not devices. We also discussed that the term Platform is often used in a different sense. To do so, I had to change the textual definition of Observer a bit. I suggest using mereotopology instead of class hierarchies here, i.e., a Sensor is mounted on (partOf, hostedBy,..) an Observer, but a Sensor is not a specific kind of Observer.I also removed all hierarchical relations that relate other classes to Device. As always, these are just proposals, nothing is set in stone. You will see later on that not having a common class for Actuators and Sensors is not perfect, but we can leave it like this for our initial discussion.

I left SamplingFeature in SOSA-core. Please read the initial rdfs:comment. IMHO, it makes a very strong and compelling case.

I left domainIncludes and rangeIncludes but changed the namespace to meta. Thanks Kerry for pointing this out.

I removed some statements that (IMHO) do not carry any semantics such as '  meta:rangeIncludes owl:Thing ;' or 'rdfs:subClassOf owl:Thing ;'.

I removed some relations that we cannot use without subclasses of Procedure or Activity. One example is the samplingStrategy relation. I am not saying that this is the right path to take but simply do so to reflect yesterday's discussion.

Overall, the implemented changes reduce SOSA-core to 9 classes and 11 relations (those will need more work in another commit).