This was to correct some errors in the ssn-minus-dolce version that we were working with on webprotege.
For this, I started from ssn as it was published by the XG at
..ssn/ssnx/ssn.owl and replaced all the dul  at loa-cnr uris to dul at
ontologydesignpatterns, changed the ssn uris to the new w3c namespace and
as / uris, added a dc:terms modified annotation.  Then deleted all the dul
axioms from the ssn file and created a new dul-alignment file for them --
and imported the new dul into that dul-alignment on ontology.

There are still a good many more thingsto tidy up, but this is the basic
starting point.