fixing issues spotted by @larsgsvensson
In Lars’ email [1] of Feb 2017 he noted some editorial issues:
1) In §5 [4] you refer to the Deutsche Nationalbibliothek (yay!).
Please don't use the URL you see in the browser. Instead use the
CMS-independent one [5].
2) There are two places in the document where references start with two
square brackets "[[". As a result there are no hyperlinks to the
(missing) references section.
3) s/converstion/conversion/ (somewhere in sec 8)
4) §8 and BP 17 say "Alternatively you can re-project your coordinates
to WGS84 Long/Lat using many available tools online." Do we want to
point to specific tools?

(1), (2)* and (3) are fixed here. (4) was fixed in an earlier edit.

* although I could only find one of these `[[skos-primer]`

[1]:
https://lists.w3.org/Archives/Public/public-sdw-wg/2017Feb/0242.html