12.2 BP5 Positional accuracy
Drop encouragement to specify the precision of the data (as the note says, that's misleading anyway)
Re-draft intended outcome in terms of user fitness-for-purpose (included summary of alternative statements here)
Added quite a large section in "how" on the various approaches.
Amended the short list of 'example(s) to be added'
'spatialResolutionAsDistance' only really applies to coverages - although it is sometimes (ab)used as a short hand for some statement about the source imagery used in the production of a vector dataset!
The INSPIRE regulation says nothing about positional accuracy, so conforming to it cannot be an example of a positional accuracy metadata statement; moved it to the more general section - but this makes it an example of a dwbp best practice. But are we OK with an example outside of the 'best practices'? There are other EC Directives that set positional accuracy requirements (e.g. for farm payments, https://marswiki.jrc.ec.europa.eu/wikicap/index.php/Positional_Accuracy - so it would be possible to create an example citing one); a better example might be IHO's S-44 which sets out accuracy requirements for hydropgrahic surveys - so survey datasets could state that they conform to S-44 (at one of four spec levels) - I've tried to create this one (although it doesn't have a URI - I've used the document URL & made up a fragment identifier)
Added a few points on on 'how to test'

Note: this page is rather long for collaborative editing; could it be broken up into sections on separate GitHub pages?