updated scope to reflect focus on _geospatial_ data
Responding to @ByronCinNZ ’s comment in [his email
(3-April)](https://lists.w3.org/Archives/Public/public-sdw-wg/2017Apr/00
09.html).

He says:

> Finally, a point about geospatial vs. other spatial in the BP doc in
general.  I do agree we need more precision in the doc in this regard.
My suggestion is, as a start, that to make it clear that where we are
speaking specifically about geospatial data we use the term
"geospatial" or similar throughout the doc.  Include the definition of
"geospatial" early on and also in the glossary.

So I’ve updated the scope to reflect our focus on geospatial data.