fixing "how to use" section (part 4)
“how to use” section completely removed; having re-read it today
(nine-months after I wrote it) I felt it largely duplicated information
that is now present in other parts of the BP document. This is in a
large part due to the restructuring and addition of much more useful
material in the BPs themselves.

As such, I felt the best thing to do was delete it after making sure
that any useful (& unduplicated) information was incorporated elsewhere.