[home page] Drop WebEx info, and link to restricted WebEx page instead
Per the WebEx Best Practices [1] page, WebEx coordinates should no longer be
shared on public pages. This includes the meeting password (which was already
behind a restricted page), but also the meeting number, and the meeting URL.

I created a dedicated page, which is only accessible to W3C members and IG
participants:
https://www.w3.org/2017/sdwig/webex.html

The home page now links to that page and no longer contain any WebEx coordinate.

[1] https://www.w3.org/2006/tools/wiki/WebExBestPractices#Keeping_WebEx_Coordinates_Secure