Update BP-exemplars.md
"Key registers" is a better description than "A-dienst". It refers to several datasets. Kadaster is not the dataset owner, but the party responsible for publishing several key registers containing Dutch reference data such as topographic data (at two different levels of detail), cadastral parcels and buildings & addresses. I'll chase the links for these.
"B-dienst" are experimental data publications that may not be maintained. I think it's better to remove these.