Consistency on "data is" rather than "data are"
In most places, the SDW BP uses "data is" rather than "data are"; only BP 6 uses "data are". 

The remaining occurrences of "data are" are in the context of e.g. "small pieces of data are" or "best practices relating to handling sensor data are" - i.e. the are is in number agreement with something other than the word "data".

I don't want to open that debate - simply to be internally consistent. Not conclusive, but searching w3.org for "data is" returns 15,400 results whereas "data are" returns 4,600.