time-prov: Drop alignment axioms that lead to deprecated property
The file time-prov.ttl provides a non-normative alignment between
OWL-Time and PROV-O.  One of the included suggestions is an alignment of
`prov:endedAtTime` with `time:inXSDDateTime` via a property chain.
Unfortunately, in the current draft of OWL-Time, `time:inXSDDateTime` is
deprecated.  A similar issue is present for `prov:startedAtTime` and
`prov:atTime`.

This patch drops alignment axioms that would lead an adopting user to
cause their PROV-O data to yield deprecated property usage under OWL
entailment.

This contribution is made only by myself, and is not being made by the
National Institute of Standards and Technology or any other
organization.

References:
* https://github.com/w3c/sdw/issues/1421
* https://www.w3.org/TR/2022/CRD-owl-time-20221115/#time:inXSDDateTime

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>